refactor: Location 객체의 필드 추가 메서드를 이미지 필드, 도보 정보 필드 추가 메서드로 구체화
